Updated: mplayer now display correctly the picture
The problem with -geometry is it seems to have a different behavior
depending on the build, the -vo used or maybe something else...

Therefore I prefer to used -vf dsize= which will resize at most at the
size given and keep aspect. Plus -vf expand= to draw a black border all
around.

The black border is useful to hide windows below the movie but also to
expand the drawing area of the subtitles. Could be very useful!