Updated: run the starting script in background
It seems my dwm-session script and the dwm process itself receive
sometimes (for some reason I don't understand) a STOP signal and freeze
my commands...

It looks related to the connectivity because as soon as I disconnect the
cable, the processes get stopped. But I'm not sure and I didn't really
investigate.

It looks better with this modification.