Fixed: closing STDOUT and STDERR with fish seems to be a bad idea
When I run gajim with >&- ^&- &, the accounts doesn't connect
automatically. But if I redirect them to a file, it works fine. So I
decided to redirect to /dev/null. This behavior is very weird but in
fact: closing an fd and redirecting it to /dev/null are two different
things.